# server/internal/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "adminconfig.calendarEventListResponse": {
                "properties": {
                    "calendarEvents": {
                        "items": {
                            "$ref": "#/components/schemas/adminconfig.calendarEventJSON"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "nextCursor": {
                        "type": "string"
                    }
                },
                "required": [
                    "calendarEvents",
                    "nextCursor"
                ],
                "type": "object"
            },
...
            "adminconfig.createCalendarEventRequest": {
                "properties": {
                    "actual": {
                        "example": 261000,
                        "type": "number"
                    },
                    "consensus": {
                        "example": 255000,
                        "type": "number"
                    },
                    "entityCodes": {
                        "example": [
                            "US"
                        ],
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "indicatorId": {
                        "example": "6b5f8e2d-4c9a-4f1e-b8d7-3a2c6e9f1b40",
                        "type": "string"
                    },
                    "name": {
                        "example": "US CPI",
                        "type": "string"
                    },
                    "previous": {
                        "example": 250000,
                        "type": "number"
                    },
                    "releasedAt": {
                        "example": "2026-10-06T12:30:00Z",
                        "type": "string"
                    },
                    "scheduledAt": {
                        "example": "2026-10-06T12:30:00Z",
                        "type": "string"
                    },
                    "sourceId": {
                        "example": "5a1c3e7b-9d2f-4c8a-b6e1-7f4d2a9c3e50",
                        "type": "string"
                    }
                },
                "required": [
                    "actual",
                    "consensus",
                    "entityCodes",
                    "indicatorId",
                    "name",
                    "previous",
                    "releasedAt",
                    "scheduledAt",
                    "sourceId"
                ],
                "type": "object"
            },
...
            "openapi.Error": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/openapi.ErrorDetail"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "Macro Terminal macro intelligence and fundamental research terminal API.",
        "title": "Macro Terminal API",
        "version": "0.1.0"
    },
...
    "paths": {
...
        "/api/v1/admin/calendar-events": {
            "get": {
                "description": "Returns stored events newest first. Pass the returned nextCursor to fetch the next page.",
                "parameters": [
                    {
                        "description": "Page size between 1 and 100. Defaults to 25.",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor returned as nextCursor by the previous page.",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.calendarEventListResponse"
                                }
                            }
                        },
                        "description": "The events and the cursor for the next page, if any."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The limit or cursor is invalid."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "List calendar events.",
                "tags": [
                    "admin"
                ]
            },
            "post": {
                "description": "Stores an event under an existing source and indicator, optionally linked to entities. Values are optional; the scheduled time is required. The same source, indicator, and scheduled time may appear only once.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminconfig.createCalendarEventRequest",
                                "summary": "body",
                                "description": "The event to store."
                            }
                        }
                    },
                    "description": "The event to store.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.calendarEventResponse"
                                }
                            }
                        },
                        "description": "The stored event."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "A field is missing or a reference is unknown."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The source, indicator, or an entity code does not exist."
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "An event with that source, indicator, and scheduled time already exists."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Create a calendar event.",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/calendar-events/{id}": {
            "patch": {
                "description": "Overwrites the event's displayed name, schedule, and figures. Provider revisions continue from the stored count.",
                "parameters": [
                    {
                        "description": "The event id.",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminconfig.updateCalendarEventRequest",
                                "summary": "body",
                                "description": "The fields to store."
                            }
                        }
                    },
                    "description": "The fields to store.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.calendarEventResponse"
                                }
                            }
                        },
                        "description": "The updated event."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The body is invalid or the scheduled time is missing."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The event does not exist."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Update a calendar event.",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/calendar-events/{id}/archive": {
            "post": {
                "description": "Hides the event from the user calendar. Archived events stay in the admin list and can be restored.",
                "parameters": [
                    {
                        "description": "The event id.",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The event is archived."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The event id is not a UUID."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The event does not exist or is already archived."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Archive a calendar event.",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/calendar-events/{id}/restore": {
            "post": {
                "description": "Returns an archived event to the user calendar.",
                "parameters": [
                    {
                        "description": "The event id.",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The event is restored."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The event id is not a UUID."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The event does not exist or is not archived."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Restore a calendar event.",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/entities": {
            "get": {
                "description": "Returns reference entities newest first. Pass the returned nextCursor to fetch the next page.",
                "parameters": [
                    {
                        "description": "Page size between 1 and 100. Defaults to 25.",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor returned as nextCursor by the previous page.",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.entityListResponse"
                                }
                            }
                        },
                        "description": "The entities and the cursor for the next page, if any."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The limit or cursor is invalid."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "List entities.",
                "tags": [
                    "admin"
                ]
            },
            "post": {
                "description": "Stores a reference entity. The code must not already exist.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminconfig.createEntityRequest",
                                "summary": "body",
                                "description": "The entity to store."
                            }
                        }
                    },
                    "description": "The entity to store.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.entityResponse"
                                }
                            }
                        },
                        "description": "The stored entity."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "A field is missing or the body is not valid JSON."
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "An entity with that code already exists."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Create an entity.",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/entity-pairs": {
            "get": {
                "description": "Returns tradable pairs newest first. Pass the returned nextCursor to fetch the next page.",
                "parameters": [
                    {
                        "description": "Page size between 1 and 100. Defaults to 25.",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor returned as nextCursor by the previous page.",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.entityPairListResponse"
                                }
                            }
                        },
                        "description": "The pairs and the cursor for the next page, if any."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The limit or cursor is invalid."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "List entity pairs.",
                "tags": [
                    "admin"
                ]
            },
            "post": {
                "description": "Stores a pair of existing entities under an explicit product symbol. Both codes must exist and differ.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminconfig.createEntityPairRequest",
                                "summary": "body",
                                "description": "The pair to store."
                            }
                        }
                    },
                    "description": "The pair to store.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminconfig.entityPairResponse"
                                }
                            }
                        },
                        "description": "The stored pair."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "A field is missing or a code is unknown."
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "An entity pair with that symbol already exists."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/openapi.Error"
                                }
                            }
                        },
                        "description": "The request could not be completed."
                    }
                },
                "summary": "Create an entity pair.",
                "tags": [
                    "admin"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Local development",
            "url": "http://localhost:8081"
        }
    ]
}